requirements.txt
torch
opencv-python
numpy<2
pillow
rembg[gpu]
diffusers
transformers
accelerate
gradio